Rename SearchBarFullScreenScreen to SearchBarScreen and SearchBarScreen to SearchBarListScreen
